Moved the notification channel creation back to the main activity because I think when I moved it was when it stopped working reliably.
I also turned off "fitsSystemWindows" in ViewDailyLog. I apparently turned it on a while ago in a commit called "fixed random bugs" but it moved everything down a bit and cut off the bottom of my screen, making it hard to read the last line of my overall notes. Turning it off moved it all back up and now it looks like a normal app again. I don't know what it was trying to do.